:octocat: Refactor comments and added tests for module

+++ src/python-module/fdpp/functions.h
@@ -14,7 +14,7 @@ int max(int x, int y) {
 }
 
 
-int min_sum_partition1(int n, int s[n], int s1, int s2) {
+int min_sum_partition1(long n, int s[n], int s1, int s2) {
 	if(n < 0)
 		return abs(s1-s2);
 	int inc = min_sum_partition1(n-1, s, s1+s[n], s2);
